chore(Intercom): move sdk loader to different file
